added something, took the generator away
Add the functionality of saving the address book to disk and restoring it from disk.
To do this, you can choose any data serialization/deserialization protocol that is convenient
for you and implement methods that will save all data to a file and load it from a file.
File: "ABook.bdata"

Add the ability to search the contents of the contact book to the user, so that all
information about one or more users can be found by a few digits of the phone number
or letters of the name, etc.
Command:
find ...
(where ... is a string(s) query? for examples:
find 063
find 044 e_7 Dou 1988)